Add script to work with OSM and saved US polygons
This patch adds a new file called osm.py which contains functions for
working with the OpenStreetMap APIs and data. The functions could be
used to obtain, convert, and save polygons from OpenStreetMap.

The patch also adds boundary polygons on the GeoJSON format for the US
and individual US states.